Move conversion.Scheme to runtime
There is only one type registry to rule them all
