+++ pkg/runtime/conversion_generator.go
@@ -27,7 +27,6 @@ import (
 	"strings"
 
 	"k8s.io/kubernetes/pkg/api/unversioned"
-	"k8s.io/kubernetes/pkg/conversion"
 	"k8s.io/kubernetes/pkg/util/sets"
 )
 
@@ -42,7 +41,7 @@ type ConversionGenerator interface {
 	AssumePrivateConversions()
 }
 
-func NewConversionGenerator(scheme *conversion.Scheme, targetPkg string) ConversionGenerator {
+func NewConversionGenerator(scheme *Scheme, targetPkg string) ConversionGenerator {
 	g := &conversionGenerator{
 		scheme: scheme,
 
@@ -66,7 +65,7 @@ func NewConversionGenerator(scheme *conversion.Scheme, targetPkg string) Convers
 var complexTypes []reflect.Kind = []reflect.Kind{reflect.Map, reflect.Ptr, reflect.Slice, reflect.Interface, reflect.Struct}
 
 type conversionGenerator struct {
-	scheme *conversion.Scheme
+	scheme *Scheme
 
 	nameFormat          string
 	generatedNamePrefix string
@@ -105,7 +104,7 @@ func (g *conversionGenerator) GenerateConversionsForType(gv unversioned.GroupVer
 	internalVersion := gv
 	internalVersion.Version = APIVersionInternal
 
-	internalObj, err := g.scheme.NewObject(internalVersion.WithKind(kind))
+	internalObj, err := g.scheme.New(internalVersion.WithKind(kind))
 	if err != nil {
 		return fmt.Errorf("cannot create an object of type %v in internal version", kind)
 	}
