+++ pkg/runtime/scheme_test.go
@@ -20,6 +20,9 @@ import (
 	"reflect"
 	"testing"
 
+	"github.com/google/gofuzz"
+	flag "github.com/spf13/pflag"
+
 	"k8s.io/kubernetes/pkg/api/unversioned"
 	"k8s.io/kubernetes/pkg/conversion"
 	"k8s.io/kubernetes/pkg/runtime"
@@ -27,29 +30,16 @@ import (
 	"k8s.io/kubernetes/pkg/util"
 )
 
-type TypeMeta struct {
-	Kind       string `json:"kind,omitempty"`
-	APIVersion string `json:"apiVersion,omitempty"`
-}
-
-// SetGroupVersionKind satisfies the ObjectKind interface for all objects that embed TypeMeta
-func (obj *TypeMeta) SetGroupVersionKind(gvk *unversioned.GroupVersionKind) {
-	obj.APIVersion, obj.Kind = gvk.ToAPIVersionAndKind()
-}
-
-// GroupVersionKind satisfies the ObjectKind interface for all objects that embed TypeMeta
-func (obj *TypeMeta) GroupVersionKind() *unversioned.GroupVersionKind {
-	return unversioned.FromAPIVersionAndKind(obj.APIVersion, obj.Kind)
-}
+var fuzzIters = flag.Int("fuzz-iters", 50, "How many fuzzing iterations to do.")
 
 type InternalSimple struct {
-	TypeMeta   `json:",inline"`
-	TestString string `json:"testString"`
+	runtime.TypeMeta `json:",inline"`
+	TestString       string `json:"testString"`
 }
 
 type ExternalSimple struct {
-	TypeMeta   `json:",inline"`
-	TestString string `json:"testString"`
+	runtime.TypeMeta `json:",inline"`
+	TestString       string `json:"testString"`
 }
 
 func (obj *InternalSimple) GetObjectKind() unversioned.ObjectKind { return &obj.TypeMeta }
@@ -135,7 +125,7 @@ func TestScheme(t *testing.T) {
 	}
 	// clearing TypeMeta is a function of the scheme, which we do not test here (ConvertToVersion
 	// does not automatically clear TypeMeta anymore).
-	simple.TypeMeta = TypeMeta{Kind: "Simple", APIVersion: externalGV.String()}
+	simple.TypeMeta = runtime.TypeMeta{Kind: "Simple", APIVersion: externalGV.String()}
 	if e, a := simple, obj3; !reflect.DeepEqual(e, a) {
 		t.Errorf("Expected:\n %#v,\n Got:\n %#v", e, a)
 	}
@@ -188,33 +178,33 @@ func TestBadJSONRejection(t *testing.T) {
 }
 
 type ExtensionA struct {
-	TypeMeta   `json:",inline"`
-	TestString string `json:"testString"`
+	runtime.TypeMeta `json:",inline"`
+	TestString       string `json:"testString"`
 }
 
 type ExtensionB struct {
-	TypeMeta   `json:",inline"`
-	TestString string `json:"testString"`
+	runtime.TypeMeta `json:",inline"`
+	TestString       string `json:"testString"`
 }
 
 type ExternalExtensionType struct {
-	TypeMeta  `json:",inline"`
-	Extension runtime.RawExtension `json:"extension"`
+	runtime.TypeMeta `json:",inline"`
+	Extension        runtime.RawExtension `json:"extension"`
 }
 
 type InternalExtensionType struct {
-	TypeMeta  `json:",inline"`
-	Extension runtime.Object `json:"extension"`
+	runtime.TypeMeta `json:",inline"`
+	Extension        runtime.Object `json:"extension"`
 }
 
 type ExternalOptionalExtensionType struct {
-	TypeMeta  `json:",inline"`
-	Extension runtime.RawExtension `json:"extension,omitempty"`
+	runtime.TypeMeta `json:",inline"`
+	Extension        runtime.RawExtension `json:"extension,omitempty"`
 }
 
 type InternalOptionalExtensionType struct {
-	TypeMeta  `json:",inline"`
-	Extension runtime.Object `json:"extension,omitempty"`
+	runtime.TypeMeta `json:",inline"`
+	Extension        runtime.Object `json:"extension,omitempty"`
 }
 
 func (obj *ExtensionA) GetObjectKind() unversioned.ObjectKind                    { return &obj.TypeMeta }
@@ -284,28 +274,28 @@ func TestExtensionMapping(t *testing.T) {
 			},
 			&InternalExtensionType{
 				Extension: &runtime.Unknown{
-					RawJSON: []byte(`{"kind":"A","apiVersion":"test.group/testExternal","testString":"foo"}`),
+					RawJSON: []byte(`{"apiVersion":"test.group/testExternal","kind":"A","testString":"foo"}`),
 				},
 			},
 			// apiVersion is set in the serialized object for easier consumption by clients
-			`{"kind":"ExtensionType","apiVersion":"` + externalGV.String() + `","extension":{"kind":"A","apiVersion":"test.group/testExternal","testString":"foo"}}
+			`{"apiVersion":"` + externalGV.String() + `","kind":"ExtensionType","extension":{"apiVersion":"test.group/testExternal","kind":"A","testString":"foo"}}
 `,
 		}, {
 			&InternalExtensionType{Extension: runtime.NewEncodable(codec, &ExtensionB{TestString: "bar"})},
 			&InternalExtensionType{
 				Extension: &runtime.Unknown{
-					RawJSON: []byte(`{"kind":"B","apiVersion":"test.group/testExternal","testString":"bar"}`),
+					RawJSON: []byte(`{"apiVersion":"test.group/testExternal","kind":"B","testString":"bar"}`),
 				},
 			},
 			// apiVersion is set in the serialized object for easier consumption by clients
-			`{"kind":"ExtensionType","apiVersion":"` + externalGV.String() + `","extension":{"kind":"B","apiVersion":"test.group/testExternal","testString":"bar"}}
+			`{"apiVersion":"` + externalGV.String() + `","kind":"ExtensionType","extension":{"apiVersion":"test.group/testExternal","kind":"B","testString":"bar"}}
 `,
 		}, {
 			&InternalExtensionType{Extension: nil},
 			&InternalExtensionType{
 				Extension: nil,
 			},
-			`{"kind":"ExtensionType","apiVersion":"` + externalGV.String() + `","extension":null}
+			`{"apiVersion":"` + externalGV.String() + `","kind":"ExtensionType","extension":null}
 `,
 		},
 	}
@@ -411,7 +401,273 @@ func TestUnversionedTypes(t *testing.T) {
 	if err != nil {
 		t.Fatal(err)
 	}
-	if string(data) != `{"kind":"InternalSimple","apiVersion":"test.group/testExternal","testString":"I'm the same"}`+"\n" {
+	if string(data) != `{"apiVersion":"test.group/testExternal","kind":"InternalSimple","testString":"I'm the same"}`+"\n" {
 		t.Errorf("unexpected data: %s", data)
+	}
+}
+
+// Test a weird version/kind embedding format.
+type MyWeirdCustomEmbeddedVersionKindField struct {
+	ID         string `json:"ID,omitempty"`
+	APIVersion string `json:"myVersionKey,omitempty"`
+	ObjectKind string `json:"myKindKey,omitempty"`
+	Z          string `json:"Z,omitempty"`
+	Y          uint64 `json:"Y,omitempty"`
+}
+
+type TestType1 struct {
+	MyWeirdCustomEmbeddedVersionKindField `json:",inline"`
+	A                                     string               `json:"A,omitempty"`
+	B                                     int                  `json:"B,omitempty"`
+	C                                     int8                 `json:"C,omitempty"`
+	D                                     int16                `json:"D,omitempty"`
+	E                                     int32                `json:"E,omitempty"`
+	F                                     int64                `json:"F,omitempty"`
+	G                                     uint                 `json:"G,omitempty"`
+	H                                     uint8                `json:"H,omitempty"`
+	I                                     uint16               `json:"I,omitempty"`
+	J                                     uint32               `json:"J,omitempty"`
+	K                                     uint64               `json:"K,omitempty"`
+	L                                     bool                 `json:"L,omitempty"`
+	M                                     map[string]int       `json:"M,omitempty"`
+	N                                     map[string]TestType2 `json:"N,omitempty"`
+	O                                     *TestType2           `json:"O,omitempty"`
+	P                                     []TestType2          `json:"Q,omitempty"`
+}
+
+type TestType2 struct {
+	A string `json:"A,omitempty"`
+	B int    `json:"B,omitempty"`
+}
+
+type ExternalTestType2 struct {
+	A string `json:"A,omitempty"`
+	B int    `json:"B,omitempty"`
+}
+type ExternalTestType1 struct {
+	MyWeirdCustomEmbeddedVersionKindField `json:",inline"`
+	A                                     string                       `json:"A,omitempty"`
+	B                                     int                          `json:"B,omitempty"`
+	C                                     int8                         `json:"C,omitempty"`
+	D                                     int16                        `json:"D,omitempty"`
+	E                                     int32                        `json:"E,omitempty"`
+	F                                     int64                        `json:"F,omitempty"`
+	G                                     uint                         `json:"G,omitempty"`
+	H                                     uint8                        `json:"H,omitempty"`
+	I                                     uint16                       `json:"I,omitempty"`
+	J                                     uint32                       `json:"J,omitempty"`
+	K                                     uint64                       `json:"K,omitempty"`
+	L                                     bool                         `json:"L,omitempty"`
+	M                                     map[string]int               `json:"M,omitempty"`
+	N                                     map[string]ExternalTestType2 `json:"N,omitempty"`
+	O                                     *ExternalTestType2           `json:"O,omitempty"`
+	P                                     []ExternalTestType2          `json:"Q,omitempty"`
+}
+
+type ExternalInternalSame struct {
+	MyWeirdCustomEmbeddedVersionKindField `json:",inline"`
+	A                                     TestType2 `json:"A,omitempty"`
+}
+
+func (obj *MyWeirdCustomEmbeddedVersionKindField) GetObjectKind() unversioned.ObjectKind { return obj }
+func (obj *MyWeirdCustomEmbeddedVersionKindField) SetGroupVersionKind(gvk *unversioned.GroupVersionKind) {
+	obj.APIVersion, obj.ObjectKind = gvk.ToAPIVersionAndKind()
+}
+func (obj *MyWeirdCustomEmbeddedVersionKindField) GroupVersionKind() *unversioned.GroupVersionKind {
+	return unversioned.FromAPIVersionAndKind(obj.APIVersion, obj.ObjectKind)
+}
+
+func (obj *ExternalInternalSame) GetObjectKind() unversioned.ObjectKind {
+	return &obj.MyWeirdCustomEmbeddedVersionKindField
+}
+
+func (obj *TestType1) GetObjectKind() unversioned.ObjectKind {
+	return &obj.MyWeirdCustomEmbeddedVersionKindField
+}
+
+func (obj *ExternalTestType1) GetObjectKind() unversioned.ObjectKind {
+	return &obj.MyWeirdCustomEmbeddedVersionKindField
+}
+
+func (obj *TestType2) GetObjectKind() unversioned.ObjectKind { return unversioned.EmptyObjectKind }
+func (obj *ExternalTestType2) GetObjectKind() unversioned.ObjectKind {
+	return unversioned.EmptyObjectKind
+}
+
+// TestObjectFuzzer can randomly populate all the above objects.
+var TestObjectFuzzer = fuzz.New().NilChance(.5).NumElements(1, 100).Funcs(
+	func(j *MyWeirdCustomEmbeddedVersionKindField, c fuzz.Continue) {
+		// We have to customize the randomization of MyWeirdCustomEmbeddedVersionKindFields because their
+		// APIVersion and Kind must remain blank in memory.
+		j.APIVersion = ""
+		j.ObjectKind = ""
+		j.ID = c.RandString()
+	},
+)
+
+// Returns a new Scheme set up with the test objects.
+func GetTestScheme() *runtime.Scheme {
+	internalGV := unversioned.GroupVersion{Version: "__internal"}
+	externalGV := unversioned.GroupVersion{Version: "v1"}
+
+	s := runtime.NewScheme()
+	// Ordinarily, we wouldn't add TestType2, but because this is a test and
+	// both types are from the same package, we need to get it into the system
+	// so that converter will match it with ExternalType2.
+	s.AddKnownTypes(internalGV, &TestType1{}, &TestType2{}, &ExternalInternalSame{})
+	s.AddKnownTypes(externalGV, &ExternalInternalSame{})
+	s.AddKnownTypeWithName(externalGV.WithKind("TestType1"), &ExternalTestType1{})
+	s.AddKnownTypeWithName(externalGV.WithKind("TestType2"), &ExternalTestType2{})
+	s.AddKnownTypeWithName(internalGV.WithKind("TestType3"), &TestType1{})
+	s.AddKnownTypeWithName(externalGV.WithKind("TestType3"), &ExternalTestType1{})
+	return s
+}
+
+func TestKnownTypes(t *testing.T) {
+	s := GetTestScheme()
+	if len(s.KnownTypes(unversioned.GroupVersion{Group: "group", Version: "v2"})) != 0 {
+		t.Errorf("should have no known types for v2")
+	}
+
+	types := s.KnownTypes(unversioned.GroupVersion{Version: "v1"})
+	for _, s := range []string{"TestType1", "TestType2", "TestType3", "ExternalInternalSame"} {
+		if _, ok := types[s]; !ok {
+			t.Errorf("missing type %q", s)
+		}
+	}
+}
+
+func TestConvertToVersion(t *testing.T) {
+	s := GetTestScheme()
+	tt := &TestType1{A: "I'm not a pointer object"}
+	other, err := s.ConvertToVersion(tt, "v1")
+	if err != nil {
+		t.Fatalf("Failure: %v", err)
+	}
+	converted, ok := other.(*ExternalTestType1)
+	if !ok {
+		t.Fatalf("Got wrong type")
+	}
+	if tt.A != converted.A {
+		t.Fatalf("Failed to convert object correctly: %#v", converted)
+	}
+}
+
+func TestMetaValues(t *testing.T) {
+	internalGV := unversioned.GroupVersion{Group: "test.group", Version: "__internal"}
+	externalGV := unversioned.GroupVersion{Group: "test.group", Version: "externalVersion"}
+
+	s := runtime.NewScheme()
+	s.AddKnownTypeWithName(internalGV.WithKind("Simple"), &InternalSimple{})
+	s.AddKnownTypeWithName(externalGV.WithKind("Simple"), &ExternalSimple{})
+
+	internalToExternalCalls := 0
+	externalToInternalCalls := 0
+
+	// Register functions to verify that scope.Meta() gets set correctly.
+	err := s.AddConversionFuncs(
+		func(in *InternalSimple, out *ExternalSimple, scope conversion.Scope) error {
+			t.Logf("internal -> external")
+			if e, a := internalGV.String(), scope.Meta().SrcVersion; e != a {
+				t.Fatalf("Expected '%v', got '%v'", e, a)
+			}
+			if e, a := externalGV.String(), scope.Meta().DestVersion; e != a {
+				t.Fatalf("Expected '%v', got '%v'", e, a)
+			}
+			scope.Convert(&in.TestString, &out.TestString, 0)
+			internalToExternalCalls++
+			return nil
+		},
+		func(in *ExternalSimple, out *InternalSimple, scope conversion.Scope) error {
+			t.Logf("external -> internal")
+			if e, a := externalGV.String(), scope.Meta().SrcVersion; e != a {
+				t.Errorf("Expected '%v', got '%v'", e, a)
+			}
+			if e, a := internalGV.String(), scope.Meta().DestVersion; e != a {
+				t.Fatalf("Expected '%v', got '%v'", e, a)
+			}
+			scope.Convert(&in.TestString, &out.TestString, 0)
+			externalToInternalCalls++
+			return nil
+		},
+	)
+	if err != nil {
+		t.Fatalf("unexpected error: %v", err)
+	}
+	simple := &InternalSimple{
+		TestString: "foo",
+	}
+
+	s.Log(t)
+
+	out, err := s.ConvertToVersion(simple, externalGV.String())
+	if err != nil {
+		t.Fatalf("unexpected error: %v", err)
+	}
+
+	internal, err := s.ConvertToVersion(out, internalGV.String())
+	if err != nil {
+		t.Fatalf("unexpected error: %v", err)
+	}
+
+	if e, a := simple, internal; !reflect.DeepEqual(e, a) {
+		t.Errorf("Expected:\n %#v,\n Got:\n %#v", e, a)
+	}
+
+	if e, a := 1, internalToExternalCalls; e != a {
+		t.Errorf("Expected %v, got %v", e, a)
+	}
+	if e, a := 1, externalToInternalCalls; e != a {
+		t.Errorf("Expected %v, got %v", e, a)
+	}
+}
+
+func TestMetaValuesUnregisteredConvert(t *testing.T) {
+	type InternalSimple struct {
+		Version    string `json:"apiVersion,omitempty"`
+		Kind       string `json:"kind,omitempty"`
+		TestString string `json:"testString"`
+	}
+	type ExternalSimple struct {
+		Version    string `json:"apiVersion,omitempty"`
+		Kind       string `json:"kind,omitempty"`
+		TestString string `json:"testString"`
+	}
+	s := runtime.NewScheme()
+	// We deliberately don't register the types.
+
+	internalToExternalCalls := 0
+
+	// Register functions to verify that scope.Meta() gets set correctly.
+	err := s.AddConversionFuncs(
+		func(in *InternalSimple, out *ExternalSimple, scope conversion.Scope) error {
+			if e, a := "unknown/unknown", scope.Meta().SrcVersion; e != a {
+				t.Fatalf("Expected '%v', got '%v'", e, a)
+			}
+			if e, a := "unknown/unknown", scope.Meta().DestVersion; e != a {
+				t.Fatalf("Expected '%v', got '%v'", e, a)
+			}
+			scope.Convert(&in.TestString, &out.TestString, 0)
+			internalToExternalCalls++
+			return nil
+		},
+	)
+	if err != nil {
+		t.Fatalf("unexpected error: %v", err)
+	}
+
+	simple := &InternalSimple{TestString: "foo"}
+	external := &ExternalSimple{}
+	err = s.Convert(simple, external)
+	if err != nil {
+		t.Fatalf("Unexpected error: %v", err)
+	}
+	if e, a := simple.TestString, external.TestString; e != a {
+		t.Errorf("Expected %v, got %v", e, a)
+	}
+
+	// Verify that our conversion handler got called.
+	if e, a := 1, internalToExternalCalls; e != a {
+		t.Errorf("Expected %v, got %v", e, a)
 	}
 }
