+++ pkg/runtime/unversioned_test.go
@@ -14,7 +14,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package conversion_test
+package runtime_test
 
 import (
 	"encoding/json"
